feat: [split-world-state-generated-adapters-003] - [Extract relation adapters, remove the exception, and prove the cleanup is behavior-preserving]
