+++ docs/internal/processes/development-guide-relevant-files.md
@@ -93,7 +93,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `api/components/schemas/data-models/FactoryName.yaml` | Reserved current-factory identifier contract | The public `FactoryName` schema owns the reserved `UNDEFINED` default-runtime identifier and must stay aligned with `/factory/~current` docs, generated artifacts, and handler semantics. |
 | `pkg/api/openapi_contract_*_test.go`, `pkg/api/generated_contract_*_test.go`, and `pkg/api/workstation_request_projection_test.go` | API contract and projection guard surfaces | OpenAPI authored-versus-bundled assertions, generated-contract round-trip coverage, and workstation-request projection regression tests should stay split into same-package concern files with shared helper owners so the backend size gate can keep applying to maintained API contract tests without moving them into another package. |
 | `pkg/api/workstation_request_projection_test.go` | Workstation dispatch projection regression surface | Dispatch-summary ownership changes should prove observable workstation-request projection behavior here, including the split between inference-attempt-owned detail and script-backed request or response fields. |
-| `pkg/factory/projections/world_state.go`, sibling `world_state_*_generated.go` production owners, `world_state_test.go`, and sibling `world_*_support_test.go` owners | Event-first world-state projection ownership seam | Projection-visible replay or diagnostics ownership changes should prove canonical event reconstruction here by asserting `InferenceAttemptsByDispatchID`, `CompletedDispatches`, and `ProviderSessions` on reconstructed world state rather than helper-presence or package-local conversion seams; keep `world_state.go` focused on reconstruction flow, reducer state, and mutation methods, and move pure generated-to-domain topology, work-item, work-content, or payload adapters into same-package sibling `world_state_*_generated.go` owners when the projection package approaches the backend size gate instead of widening a separate helper package. |
+| `pkg/factory/projections/world_state.go`, sibling `world_state_*_generated.go` production owners, `world_state_test.go`, and sibling `world_*_support_test.go` owners | Event-first world-state projection ownership seam | Projection-visible replay or diagnostics ownership changes should prove canonical event reconstruction here by asserting `InferenceAttemptsByDispatchID`, `CompletedDispatches`, and `ProviderSessions` on reconstructed world state rather than helper-presence or package-local conversion seams; keep `world_state.go` focused on reconstruction flow, reducer state, and mutation methods, and move pure generated-to-domain topology, work-item, work-content, relation, or payload adapters into same-package sibling `world_state_*_generated.go` owners when the projection package approaches the backend size gate instead of widening a separate helper package. |
 | `pkg/api/testdata/canonical-event-vocabulary-stream.json` | OpenAPI vocabulary fixture | Canonical bundled-contract fixtures for event payload validation live here and must be updated alongside public schema field renames. |
 | `pkg/factory/event_history.go` | Canonical event fanout owner | Live factory event subscriptions are owned here; cancellation-safe stream shutdown and non-blocking fanout behavior should stay centralized instead of being reimplemented in runtime or API layers. |
 | `pkg/factory/event_history_test.go` plus sibling `event_history_*.go` test owners | Event-history generated payload and canonical stream regression seam | Keep initial-structure payload assertions, script-boundary stream coverage, and chaining-trace lineage regressions in same-package sibling test owners so the emitted event contract stays reviewable by concern without moving helpers into a separate test package. |
